requirements.txt
pillow==11.3.0
beautifulsoup4==4.13.4
